refactor-test: удалил ненужный файл и переместил мудуль сокета в другую директорию
