--- test/stuff/conf/router0.json
@@ -1,16 +1,0 @@
-[
-    {
-        "name": "index",
-        "verb": "GET",
-        "expr": "/"
-    },
-    {
-        "name": "static",
-        "verb": "GET",
-        "expr": "/static/",
-        "opts": {
-            "nocase": true,
-            "noend": true
-        }
-    }
-]
